backend/ai/app/requirements.txt
# pdf2image==1.17.0
pinecone-client
python-dotenv
google.generativeai
Flask
#!sudo apt-get install -y poppler-utils
